organized the folders, files and configs
