+++ Makefile
@@ -1,29 +1,73 @@
+#* *************************************************************************** *#
+#*                                                                             *#
+#*  $$\      $$\           $$\                                                 *#
+#*  $$ | $\  $$ |          $$ |                                                *#
+#*  $$ |$$$\ $$ | $$$$$$\  $$$$$$$\   $$$$$$$\  $$$$$$\   $$$$$$\ $$\    $$\   *#
+#*  $$ $$ $$\$$ |$$  __$$\ $$  __$$\ $$  _____|$$  __$$\ $$  __$$\\$$\  $$  |  *#
+#*  $$$$  _$$$$ |$$$$$$$$ |$$ |  $$ |\$$$$$$\  $$$$$$$$ |$$ |  \__|\$$\$$  /   *#
+#*  $$$  / \$$$ |$$   ____|$$ |  $$ | \____$$\ $$   ____|$$ |       \$$$  /    *#
+#*  $$  /   \$$ |\$$$$$$$\ $$$$$$$  |$$$$$$$  |\$$$$$$$\ $$ |        \$  /     *#
+#*  \__/     \__| \_______|\_______/ \_______/  \_______|\__|         \_/      *#
+#*                                                                             *#
+#*   By: Panu Kangas, Ryan Boudwin, Patrik Lång                                *#
+#*                                                                             *#
+#* *************************************************************************** *#
+
 NAME = webserv
 
 COMP = c++
 
 CFLAGS = -Wall -Wextra -Werror -std=c++20
 
-SRC = main.cpp ConnectionHandler.cpp ResponseHandler.cpp ConfigurationHandler.cpp CgiHandler.cpp \
-signalHandling.cpp requestParsing.cpp Logger.cpp URLhandler.cpp
+RED = \033[31m
+GREEN = \033[32m
+RESET = \033[0m
 
-OBJ = $(SRC:.cpp=.o)
+# Paths
+SRCS_PATH = srcs/
+INCLUDES_PATH = includes/
+OBJS_PATH = objs/
 
-.PHONY: all re clean fclean
+# Source Files
+SRCS = $(addprefix $(SRCS_PATH), \
+		main.cpp \
+		ConnectionHandler.cpp \
+		ResponseHandler.cpp \
+		ConfigurationHandler.cpp \
+		CgiHandler.cpp \
+		signalHandling.cpp \
+		requestParsing.cpp \
+		Logger.cpp \
+		URLhandler.cpp)
 
+# Object Files
+OBJS = $(addprefix $(OBJS_PATH), $(notdir $(SRCS:.cpp=.o)))
+
+# Build Target
 all: $(NAME)
 
-$(NAME): $(OBJ)
-	$(COMP) $(CFLAGS) $(OBJ) -o $(NAME)
+$(NAME): $(OBJS_PATH) $(OBJS)
+	@$(COMP) $(CFLAGS) -I$(INCLUDES_PATH) $(OBJS) -o $(NAME)
+	@echo "$(GREEN)$(NAME) creation successfull!$(RESET)"
 
-%.o: %.cpp
-	$(COMP) $(CFLAGS) -o $@ -c $<
+# Compile .cpp into .o files
+$(OBJS_PATH)%.o: $(SRCS_PATH)%.cpp | $(OBJS_PATH)
+	@$(COMP) $(CFLAGS) -I$(INCLUDES_PATH) -c $< -o $@
 
+# Create obj directory if it doesn’t exist
+$(OBJS_PATH):
+	@mkdir -p $(OBJS_PATH)
+
+# Cleanup
 clean:
-	rm -f $(OBJ)
+	@rm -rf $(OBJS_PATH)
+	@echo "$(RED)Object files removed!$(RESET)"
 
 fclean: clean
-	rm -f $(NAME)
-	rm -f logfile.log
+	@rm -rf $(NAME)
+	@rm -rf logfile.log
+	@echo "$(RED)Full cleanup successfull!$(RESET)"
 
 re: fclean all
+
+.PHONY: all clean fclean re
