+++ configurations/deletetest.conf
@@ -1,19 +1,33 @@
+# listen 				- What port to listen (8000 - 8999)
+# host 					- What IP address
+# server_name 			- example.com www.example.com
+# max_client_body_size 	- the max size of a clients body
+# root 					- The path to the location (mandatory)
+# methods 				- Allowed methods at in that location
+# cgi_path_php/python 	- The path for cgi interpreter (change depending on the system)
+# dir_listing 			- If the directory listing(autoindex) is on/off
+# upload_dir 			- A specific upload directory, has to be inside home(/) (optional)
+# return 				- We allow 307 (Temporary redirect), inside a location, inside home(/) (optional)
+# index 				- Default index page
+
+# All above mentioned variables end with a ';' (server and location does not)
+
+# Server Configuration
+
 server
 {
-	listen 8080;  # Listening to port 8080;
+	listen 8080;
 	host 127.0.0.1;
-	server_name test.com www.test.com;
-	max_client_body_size 5M; # Limit request body size
+	server_name deletetest.com www.deletetest.com;
+	max_client_body_size 10000000;
 
-	# HTTP to HTTPS redirection
-    return 301 https://$host$request_uri;  # Redirect to HTTPS version of the URL
-
-	# Error pages
+	# Custom error pages
 	error_page 400 /home/error/400.html;
 	error_page 403 /home/error/403.html;
 	error_page 404 /home/error/404.html;
 	error_page 405 /home/error/405.html;
 	error_page 500 /home/error/500.html;
+	error_page 503 /home/error/503.html;
 
 	# Default index
 	index index.html;
@@ -21,40 +35,41 @@ server
 	# Routes
 	location /
 	{
-		root home; # Location of home directory
-		methods GET DELETE; # Allowed methods
-		dir_listing on; # Disables directory listing
-	}
-	location /no_del_dir
-	{
-		root home/no_del_dir;
-		methods GET POST;
+		root home;
+		methods GET DELETE;
 		dir_listing on;
-	}
-	location /no_get
-	{
-		root home/no_get;
-		methods POST;
-		dir_listing on;
-	}
-	location /images/
-	{
-		root home/images; # Location for image files
-		methods GET POST DELETE; # Allowed methods
-		dir_listing off; # Disables directory listing
-	}
-	location /uploads/
-	{
-		root home/uploads; # Location for uploading files
-		methods GET POST DELETE; # Allowed methods
-		upload_dir /home/uploads; #Upload destination
-		dir_listing on; # Disables directory listing
-	}
-	location /cgi/
-	{
-		root home/cgi; # Location for .php scripts
-		methods GET POST; # Allowed methods
-		cgi_path /usr/bin; # Path for cgi interpreter (change depending on the system)
-		dir_listing on; # Disables directory listing
+
+		location /no_del_dir
+		{
+			root home/no_del_dir;
+			methods GET POST;
+			dir_listing on;
+		}
+		location /no_get
+		{
+			root home/no_get;
+			methods POST;
+			dir_listing on;
+		}
+		location /images/
+		{
+			root home/images;
+			methods GET POST DELETE;
+			dir_listing off;
+		}
+		location /uploads/
+		{
+			root home/uploads;
+			methods GET POST DELETE;
+			upload_dir /home/uploads;
+			dir_listing on;
+		}
+		location /cgi/
+		{
+			root home/cgi;
+			methods GET POST;
+			cgi_path /usr/bin;
+			dir_listing on;
+		}
 	}
 }
