+++ configurations/panuHome.conf
@@ -5,9 +5,6 @@ server
 	server_name test.com www.test.com;
 	max_client_body_size 500000; # Limit request body size
 
-	# HTTP to HTTPS redirection
-    return 301 https://$host$request_uri;  # Redirect to HTTPS version of the URL
-
 	# Error pages
 	error_page 400 /home/error/400.html;
 	error_page 403 /home/error/403.html;
