+++ configurations/simple.conf
@@ -1,7 +1,23 @@
+# listen 				- What port to listen (8000 - 8999)
+# host 					- What IP address
+# server_name 			- example.com www.example.com
+# max_client_body_size 	- the max size of a clients body
+# root 					- The path to the location (mandatory)
+# methods 				- Allowed methods at in that location
+# cgi_path_php/python 	- The path for cgi interpreter (change depending on the system)
+# dir_listing 			- If the directory listing(autoindex) is on/off
+# upload_dir 			- A specific upload directory, has to be inside home(/) (optional)
+# return 				- We allow 307 (Temporary redirect), inside a location, inside home(/) (optional)
+# index 				- Default index page
+
+# All above mentioned variables end with a ';' (server and location does not)
+
+# Server Configuration
+
 server
 {
 	listen 8081;
-	host 127.0.0.73333333333333333;
+	host 127.0.0.1;
 	server_name simple.com www.simple.com;
 	max_client_body_size 5000;
 
@@ -17,9 +33,9 @@ server
 server
 {
 	listen 8082;
-	host 127.0.0.8;
+	host 127.0.0.2;
 	server_name simple.com www.simple.com;
-	max_client_body_size 3000;
+	max_client_body_size 5000;
 
 	# Default index
 	index index.html;
@@ -33,9 +49,9 @@ server
 server
 {
 	listen 8083;
-	host 127.0.0.9;
+	host 127.0.0.3;
 	server_name simple.com www.simple.com;
-	max_client_body_size 50000;
+	max_client_body_size 5000;
 
 	# Default index
 	index index.html;
