+++ configurations/test.conf
@@ -1,9 +1,25 @@
+# listen 				- What port to listen (8000 - 8999)
+# host 					- What IP address
+# server_name 			- example.com www.example.com
+# max_client_body_size 	- the max size of a clients body
+# root 					- The path to the location (mandatory)
+# methods 				- Allowed methods at in that location
+# cgi_path_php/python 	- The path for cgi interpreter (change depending on the system)
+# dir_listing 			- If the directory listing(autoindex) is on/off
+# upload_dir 			- A specific upload directory, has to be inside home(/) (optional)
+# return 				- We allow 307 (Temporary redirect), inside a location, inside home(/) (optional)
+# index 				- Default index page
+
+# All above mentioned variables end with a ';' (server and location does not)
+
+# Server Configuration
+
 server
 {
-	listen 8080; # port
-	host 127.0.0.1; # ip address
-	server_name test.com www.test.com; # server name
-	max_client_body_size 10000000; # max size for client request body
+	listen 8080;
+	host 127.0.0.1;
+	server_name test.com www.test.com;
+	max_client_body_size 10000000;
 
 	# Custom error pages
 	error_page 400 /home/error/400.html;
@@ -18,6 +34,7 @@ server
 	error_page 431 /home/error/431.html;
 	error_page 500 /home/error/500.html;
 	error_page 501 /home/error/501.html;
+	error_page 503 /home/error/503.html;
 	error_page 505 /home/error/505.html;
 
 	# Default index
@@ -26,33 +43,33 @@ server
 	# Routes
 	location /
 	{
-		root home; # Location of home directory
-		methods GET; # Allowed methods
+		root home;
+		methods GET;
 		cgi_path_php /usr/bin;
 		cgi_path_python /usr/bin;
-		dir_listing off; # Disables/enables directory listing
+		dir_listing off; 
 
 		location /images/
 		{
-			root home/images; # Location of image directory
-			methods GET POST DELETE; # Allowed methods
-			upload_dir home/images/uploads/; # upload directory, has to be inside home(/) (optional)
-			dir_listing off; # Disables/enables directory listing
+			root home/images;
+			methods GET POST DELETE;
+			upload_dir home/images/uploads/;
+			dir_listing off; 
 
 			location /images/uploads/
 			{
-				root home/images/uploads; # Location of uploads directory
-				methods GET POST DELETE; # Allowed methods
-				dir_listing off; # Disables/enables directory listing
+				root home/images/uploads;
+				methods GET POST DELETE;
+				dir_listing off; 
 			}
 		}
 		location /cgi/
 		{
-			root home/cgi;  # Location of cgi directory
-			methods GET POST; # Allowed methods
+			root home/cgi;
+			methods GET POST;
 			cgi_path_php /usr/bin;
-			cgi_path_python /usr/bin; # Path for cgi interpreter (change depending on the system)
-			dir_listing off; # Disables/enables directory listing
+			cgi_path_python /usr/bin;
+			dir_listing off;
 		}
 	}
 }
