+++ configurations/web.conf
@@ -1,13 +1,27 @@
+# listen 				- What port to listen (8000 - 8999)
+# host 					- What IP address
+# server_name 			- example.com www.example.com
+# max_client_body_size 	- the max size of a clients body
+# root 					- The path to the location (mandatory)
+# methods 				- Allowed methods at in that location
+# cgi_path_php/python 	- The path for cgi interpreter (change depending on the system)
+# dir_listing 			- If the directory listing(autoindex) is on/off
+# upload_dir 			- A specific upload directory, has to be inside home(/) (optional)
+# return 				- We allow 307 (Temporary redirect), inside a location, inside home(/) (optional)
+# index 				- Default index page
+
+# All above mentioned variables end with a ';' (server and location does not)
+
 # Server Configuration
 
 server
 {
-    listen 8080;
+    listen 8001;
     host 127.0.0.1;
     server_name default.com www.default.com;
     max_client_body_size 30000000;
 
-   # Error pages
+   # Custom error pages
 	error_page 400 /home/error/400.html;
 	error_page 403 /home/error/403.html;
 	error_page 404 /home/error/404.html;
@@ -20,12 +34,13 @@ server
 	error_page 431 /home/error/431.html;
 	error_page 500 /home/error/500.html;
 	error_page 501 /home/error/501.html;
+	error_page 503 /home/error/503.html;
 	error_page 505 /home/error/505.html;
 
     # Default index
     index index.html;
 
-    # Routes for different directories
+    # Routes
     location /
     {
         root home;
@@ -50,12 +65,12 @@ server
 
 server
 {
-	listen 8081;
+	listen 8002;
 	host 127.0.0.2;
 	server_name dads.com www.dads.com;
 	max_client_body_size 30000000;
 
-	# Error pages
+	# Custom error pages
 	error_page 400 /home/error/400.html;
 	error_page 403 /home/error/403.html;
 	error_page 404 /home/error/404.html;
@@ -68,12 +83,13 @@ server
 	error_page 431 /home/error/431.html;
 	error_page 500 /home/error/500.html;
 	error_page 501 /home/error/501.html;
+	error_page 503 /home/error/503.html;
 	error_page 505 /home/error/505.html;
 
 	# Default index
 	index index.html;
 
-	# Routes for different sections
+	# Routes
 	location /
 	{
 		root home;
@@ -97,12 +113,12 @@ server
 
 server
 {
-	listen 8069;
+	listen 8003;
 	host 127.0.0.3;
 	server_name pat.com www.pat.com;
 	max_client_body_size 50000000;
 
-	# Error pages
+	# Custom error pages
 	error_page 400 /home/error/400.html;
 	error_page 403 /home/error/403.html;
 	error_page 404 /home/error/404.html;
@@ -115,6 +131,7 @@ server
 	error_page 431 /home/error/431.html;
 	error_page 500 /home/error/500.html;
 	error_page 501 /home/error/501.html;
+	error_page 503 /home/error/503.html;
 	error_page 505 /home/error/505.html;
 
 	# Default index
