+++ includes/ConfigurationHandler.hpp
@@ -13,16 +13,15 @@
 /*                                                                             */
 /* *************************************************************************** */
 
-#ifndef CONFIGURATIONHANDLER_HPP
-# define CONFIGURATIONHANDLER_HPP
+#pragma once
 
-# include <vector>
-# include <map>
-# include <regex>
+#include <vector>
+#include <map>
+#include <regex>
 
-# define G_CGI_PATH_PHP	"/usr/bin"
-# define G_CGI_PATH_PYTHON "/usr/bin"
-# define G_METHOD	"GET"
+#define G_CGI_PATH_PHP	"/usr/bin"
+#define G_CGI_PATH_PYTHON "/usr/bin"
+#define G_METHOD	"GET"
 
 enum dirListStates
 {
@@ -33,14 +32,14 @@ enum dirListStates
 
 struct locationBlock
 {
-	std::string					m_root;						// root of each location block in server configuration
-	int							m_reDirectStatusCode = 0;	// redirect, status code
-	std::string					m_reDirectLocation;			// redirect, /new_location
-	std::string					m_methods; 					// e.g. GET POST DELETE
-	std::string					m_uploadDir;				// specified upload directory
-	std::string					m_cgiPathPHP;				// path to cgi interpreter PHP, changes depending on system
-	std::string					m_cgiPathPython;			// path to cgi interpreter Python, changes depending on system
-	enum dirListStates			m_dirListing; 				// directory listing ON or OFF
+	std::string			m_root;						// root of each location block in server configuration
+	int					m_reDirectStatusCode = 0;	// redirect, status code
+	std::string			m_reDirectLocation;			// redirect, /new_location
+	std::string			m_methods; 					// e.g. GET POST DELETE
+	std::string			m_uploadDir;				// specified upload directory
+	std::string			m_cgiPathPHP;				// path to cgi interpreter PHP, changes depending on system
+	std::string			m_cgiPathPython;			// path to cgi interpreter Python, changes depending on system
+	enum dirListStates	m_dirListing; 				// directory listing ON or OFF
 
 	locationBlock() : m_dirListing(UNSET)
 	{};
@@ -70,41 +69,39 @@ class ConfigurationHandler
 		ConfigurationHandler(std::vector<std::string>);
 		~ConfigurationHandler() {};
 
-		void		defaultSettings();
+		void	defaultSettings();
 
-		std::string				getHost() const;
-		std::string				getPort() const;
-		std::string				getIndex() const;
-		unsigned int			getMCBSize() const;
-		std::string				getNames() const;
-		std::string				getInheritedMethods(std::string) const;
-		enum dirListStates		getInheritedDirListing(std::string) const;
-		std::string				getInheritedCgiPathPHP(std::string) const;
-		std::string				getInheritedCgiPathPython(std::string) const;
-		std::string				getRoot(std::string) const;
-		int						getRedirectStatusCode(std::string) const;
-		std::string				getRedirectLocation(std::string) const;
-		std::string				getUploadDir(std::string) const;
-		std::string				getMethods(std::string) const;
-		enum dirListStates		getDirListing(std::string) const;
-		std::string				getCgiPathPHP(std::string) const;
-		std::string				getCgiPathPython(std::string) const;
-		std::string				getErrorPages(unsigned int) const;
-		std::string				getDefaultErrorPages(unsigned int) const;
-		int						getErrorStatus() const;
+		std::string			getHost() const;
+		std::string			getPort() const;
+		std::string			getIndex() const;
+		unsigned int		getMCBSize() const;
+		std::string			getNames() const;
+		std::string			getInheritedMethods(std::string) const;
+		enum dirListStates	getInheritedDirListing(std::string) const;
+		std::string			getInheritedCgiPathPHP(std::string) const;
+		std::string			getInheritedCgiPathPython(std::string) const;
+		std::string			getRoot(std::string) const;
+		int					getRedirectStatusCode(std::string) const;
+		std::string			getRedirectLocation(std::string) const;
+		std::string			getUploadDir(std::string) const;
+		std::string			getMethods(std::string) const;
+		enum dirListStates	getDirListing(std::string) const;
+		std::string			getCgiPathPHP(std::string) const;
+		std::string			getCgiPathPython(std::string) const;
+		std::string			getErrorPages(unsigned int) const;
+		std::string			getDefaultErrorPages(unsigned int) const;
+		int					getErrorStatus() const;
 
-		bool					checkLocationBlocksRoot(locationBlock &);
-		bool					requiredSettings();
+		bool	checkLocationBlocksRoot(locationBlock &);
+		bool	requiredSettings();
 
-		void					printSettings();
+		bool	isLocationConfigured(std::string);
+		bool	isRedirectSet(std::string);
+		bool	isUploadDirSet(std::string);
 
-		bool					isRedirectSet(std::string);
-		bool					isUploadDirSet(std::string);
-		bool					isLocationConfigured(std::string);
+		void	printSettings();
 };
 
 std::string	fileNameCheck(char *);
 void		readFile(const std::string&, std::vector<std::string>&);
 void		extractServerBlocks(std::multimap<std::string, ConfigurationHandler>&, std::vector<std::string>&);
-
-#endif
