+++ includes/Logger.hpp
@@ -1,0 +1,44 @@
+/* *************************************************************************** */
+/*                                                                             */
+/*  $$\      $$\           $$\                                                 */
+/*  $$ | $\  $$ |          $$ |                                                */
+/*  $$ |$$$\ $$ | $$$$$$\  $$$$$$$\   $$$$$$$\  $$$$$$\   $$$$$$\ $$\    $$\   */
+/*  $$ $$ $$\$$ |$$  __$$\ $$  __$$\ $$  _____|$$  __$$\ $$  __$$\\$$\  $$  |  */
+/*  $$$$  _$$$$ |$$$$$$$$ |$$ |  $$ |\$$$$$$\  $$$$$$$$ |$$ |  \__|\$$\$$  /   */
+/*  $$$  / \$$$ |$$   ____|$$ |  $$ | \____$$\ $$   ____|$$ |       \$$$  /    */
+/*  $$  /   \$$ |\$$$$$$$\ $$$$$$$  |$$$$$$$  |\$$$$$$$\ $$ |        \$  /     */
+/*  \__/     \__| \_______|\_______/ \_______/  \_______|\__|         \_/      */
+/*                                                                             */
+/*   By: Panu Kangas, Ryan Boudwin, Patrik Lång                                */
+/*                                                                             */
+/* *************************************************************************** */
+
+#pragma once
+
+#include <sstream>
+#include <ctime>
+#include <fstream>
+#include <iostream>
+
+enum LogStatus
+{
+	INFO,
+	ERROR,
+	DEBUG,
+	WARNING
+};
+
+class Logger
+{
+	private:
+		std::ofstream logFile;
+
+	public:
+		Logger(const std::string &);
+		~Logger();
+
+		void	webservLog(LogStatus, const std::string &, bool);
+		void	closeLogFileStream();
+};
+
+extern Logger webservLog;
