+++ includes/URLhandler.hpp
@@ -12,3 +12,22 @@
 /*   By: Panu Kangas, Ryan Boudwin, Patrik Lång                                */
 /*                                                                             */
 /* *************************************************************************** */
+
+#pragma once
+
+#include <iostream>
+#include <unordered_map>
+
+class URLhandler
+{
+	private:
+		std::unordered_map<std::string, std::string> m_decodeMap;
+		std::unordered_map<char, std::string> m_encodeMap;
+
+	public:
+		URLhandler();
+		~URLhandler();
+
+		void decode(std::string &str);
+		void encode(std::string &str);
+};
