+++ srcs/ResponseHandler.cpp
@@ -1,3 +1,18 @@
+/* *************************************************************************** */
+/*                                                                             */
+/*  $$\      $$\           $$\                                                 */
+/*  $$ | $\  $$ |          $$ |                                                */
+/*  $$ |$$$\ $$ | $$$$$$\  $$$$$$$\   $$$$$$$\  $$$$$$\   $$$$$$\ $$\    $$\   */
+/*  $$ $$ $$\$$ |$$  __$$\ $$  __$$\ $$  _____|$$  __$$\ $$  __$$\\$$\  $$  |  */
+/*  $$$$  _$$$$ |$$$$$$$$ |$$ |  $$ |\$$$$$$\  $$$$$$$$ |$$ |  \__|\$$\$$  /   */
+/*  $$$  / \$$$ |$$   ____|$$ |  $$ | \____$$\ $$   ____|$$ |       \$$$  /    */
+/*  $$  /   \$$ |\$$$$$$$\ $$$$$$$  |$$$$$$$  |\$$$$$$$\ $$ |        \$  /     */
+/*  \__/     \__| \_______|\_______/ \_______/  \_______|\__|         \_/      */
+/*                                                                             */
+/*   By: Panu Kangas, Ryan Boudwin, Patrik Lång                                */
+/*                                                                             */
+/* *************************************************************************** */
+
 #include "ResponseHandler.hpp"
 #include "ConnectionHandler.hpp"
 #include "Structs.hpp"
